app/openai_client.py
import os
from dotenv import load_dotenv
from openai import OpenAI

load_dotenv()

client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))

SYSTEM_PROMPT = """
You are the admissions assistant of Colegio Montebello (Quito, Ecuador).

Your goal:
- Welcome and guide parents who are interested in the school.
- Ask a few polite questions to understand their needs:
  - Parent name
  - Student name
  - Student age or grade they are applying for
  - Contact phone
  - Contact email
  - Preferred day/time to visit the school
- Explain briefly how school tours work at Montebello (in general terms).
- Once you have enough information, clearly confirm that a tour request has been registered,
  and reassure them that the admissions team will follow up.

Important:
- Be warm, professional, short and clear.
- If some key information is missing (like phone or grade), politely ask for it.
- Do NOT invent exact prices or legal promises. If asked about costs, speak generally and say that
  exact details are provided by the admissions team.
- If the user speaks Spanish, answer in Spanish. If they start in English, answer in English.
"""


def chat_with_openai(user_message: str) -> str:
    """
    Simple one-turn chat: we send the user's latest message plus the system prompt.
    This is enough for a prototype.
    """
    response = client.chat.completions.create(
        model="gpt-3.5-turbo",  # or whichever model you prefer/are allowed to use
        messages=[
            {"role": "system", "content": SYSTEM_PROMPT},
            {"role": "user", "content": user_message},
        ],
    )

    return response.choices[0].message.content
